feat: trying to make the local save load faster for the screen and refactor
