+++ app/src/main/java/com/github/se/assocify/model/CurrentUser.kt
@@ -1,7 +1,5 @@
 package com.github.se.assocify.model
 
-import com.github.se.assocify.model.entities.Theme
-
 object CurrentUser {
   var userUid: String? = null
   var associationUid: String? = null
