+++ app/src/main/java/com/github/se/assocify/model/entities/UserPreference.kt
@@ -27,15 +27,15 @@ enum class Theme {
   DARK,
   SYSTEM;
 
-    companion object {
-        fun fromString(value: String?): Theme {
-            return when (value) {
-                "LIGHT" -> LIGHT
-                "DARK" -> DARK
-                else -> SYSTEM
-            }
-        }
+  companion object {
+    fun fromString(value: String?): Theme {
+      return when (value) {
+        "LIGHT" -> LIGHT
+        "DARK" -> DARK
+        else -> SYSTEM
+      }
     }
+  }
 }
 
 /** All the languages supported by the application */
