+++ app/src/main/java/com/github/se/assocify/navigation/NavigationGraph.kt
@@ -10,7 +10,7 @@ import com.github.se.assocify.model.database.EventAPI
 import com.github.se.assocify.model.database.ReceiptAPI
 import com.github.se.assocify.model.database.TaskAPI
 import com.github.se.assocify.model.database.UserAPI
-import com.github.se.assocify.model.localsave.LoginSave
+import com.github.se.assocify.model.localsave.LocalSave
 import com.github.se.assocify.ui.screens.chat.chatGraph
 import com.github.se.assocify.ui.screens.createAssociation.createAssociationGraph
 import com.github.se.assocify.ui.screens.event.eventGraph
@@ -33,7 +33,7 @@ fun NavGraphBuilder.mainNavGraph(
     accountingCategoriesAPI: AccountingCategoryAPI,
     accountingSubCategoryAPI: AccountingSubCategoryAPI,
     appThemeViewModel: ThemeViewModel,
-    localSave : LoginSave
+    localSave: LocalSave
 ) {
   homeGraph(navActions)
   treasuryGraph(
@@ -45,7 +45,14 @@ fun NavGraphBuilder.mainNavGraph(
       accountingSubCategoryAPI)
   eventGraph(navActions, eventAPI, taskAPI)
   chatGraph(navActions)
-  profileGraph(navActions, userAPI, associationAPI, accountingCategoriesAPI, eventAPI, appThemeViewModel, localSave)
+  profileGraph(
+      navActions,
+      userAPI,
+      associationAPI,
+      accountingCategoriesAPI,
+      eventAPI,
+      appThemeViewModel,
+      localSave)
   loginGraph(navActions, userAPI)
   selectAssociationGraph(navActions, userAPI, associationAPI)
   createAssociationGraph(navActions, userAPI, associationAPI)
