+++ app/src/main/java/com/github/se/assocify/ui/screens/profile/ProfileGraph.kt
@@ -7,7 +7,7 @@ import com.github.se.assocify.model.database.AccountingCategoryAPI
 import com.github.se.assocify.model.database.AssociationAPI
 import com.github.se.assocify.model.database.EventAPI
 import com.github.se.assocify.model.database.UserAPI
-import com.github.se.assocify.model.localsave.LoginSave
+import com.github.se.assocify.model.localsave.LocalSave
 import com.github.se.assocify.navigation.Destination
 import com.github.se.assocify.navigation.NavigationActions
 import com.github.se.assocify.ui.screens.profile.events.profileEventsGraph
@@ -22,8 +22,8 @@ fun NavGraphBuilder.profileGraph(
     associationAPI: AssociationAPI,
     accountingCategoryAPI: AccountingCategoryAPI,
     eventAPI: EventAPI,
-    appThemeVM : ThemeViewModel,
-    localSave : LoginSave
+    appThemeVM: ThemeViewModel,
+    localSave: LocalSave
 ) {
   composable(
       route = Destination.Profile.route,
