+++ app/src/main/java/com/github/se/assocify/ui/screens/profile/preferences/ProfilePreferencesScreen.kt
@@ -32,7 +32,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.unit.dp
 import com.github.se.assocify.model.entities.Theme
-import com.github.se.assocify.model.localsave.LoginSave
+import com.github.se.assocify.model.localsave.LocalSave
 import com.github.se.assocify.navigation.NavigationActions
 import com.github.se.assocify.ui.composables.DropdownOption
 import com.github.se.assocify.ui.composables.DropdownWithSetOptions
@@ -48,10 +48,14 @@ import com.github.se.assocify.ui.theme.ThemeViewModel
  */
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun ProfilePreferencesScreen(navActions: NavigationActions, appThemeViewModel: ThemeViewModel, localSave : LoginSave) {
+fun ProfilePreferencesScreen(
+    navActions: NavigationActions,
+    appThemeViewModel: ThemeViewModel,
+    localSave: LocalSave
+) {
   // temporary values for the theme, text size, language and currency, waiting for the viewmodel
   val themeOptions = listOf(Theme.LIGHT, Theme.DARK, Theme.SYSTEM)
-    val currentTheme by appThemeViewModel.theme.collectAsState()
+  val currentTheme by appThemeViewModel.theme.collectAsState()
 
   var themeSelectedIndex by remember { mutableIntStateOf(themeOptions.indexOf(currentTheme)) }
 
@@ -91,16 +95,15 @@ fun ProfilePreferencesScreen(navActions: NavigationActions, appThemeViewModel: T
                           SegmentedButtonDefaults.itemShape(
                               index = index, count = themeOptions.size),
                       onClick = {
-                          if ( index != themeSelectedIndex){
-                              themeSelectedIndex = index
-                              appThemeViewModel.setTheme(label)
-                              localSave.saveTheme()
-
-                          }},
-                      selected = index == themeSelectedIndex
-                  ) {
-                    Text(label.name)
-                  }
+                        if (index != themeSelectedIndex) {
+                          themeSelectedIndex = index
+                          appThemeViewModel.setTheme(label)
+                          localSave.saveTheme()
+                        }
+                      },
+                      selected = index == themeSelectedIndex) {
+                        Text(label.name)
+                      }
                 }
               }
